+++ lib/preferences/metric_spec.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:pref/pref.dart';
 import 'package:syncfusion_flutter_charts/charts.dart' as charts;
+import '../preferences/sport_spec.dart';
 import '../ui/models/row_configuration.dart';
 import '../utils/constants.dart';
 import '../utils/display.dart';
@@ -13,20 +14,10 @@ double decimalRound(double value, {int precision = 100}) {
 }
 
 const targetHrShortTitle = "Target HR";
-const slowSpeedPostfix = " Speed (kmh) Considered Too Slow to Display";
-const slowSpeedTagPrefix = "slow_speed_";
 
 class MetricSpec {
   static const thresholdCapital = " Threshold ";
   static const zonesCapital = " Zones (list of % of threshold)";
-  static const paddleSport = "Paddle";
-  static const sportPrefixes = [
-    ActivityType.ride,
-    ActivityType.run,
-    paddleSport,
-    ActivityType.swim,
-    ActivityType.elliptical,
-  ];
   static const thresholdPrefix = "threshold_";
   static const zonesPostfix = "_zones";
   static const metrics = ["power", "speed", "cadence", "hr"];
@@ -41,14 +32,6 @@ class MetricSpec {
   static const zoneIndexDisplayDefault = false;
   static const veryOldZoneBoundaries = "55,75,90,105,120,150";
 
-  static final slowSpeeds = {
-    ActivityType.ride: 4.0,
-    ActivityType.run: 2.0,
-    paddleSport: 1.0,
-    ActivityType.swim: 0.5,
-    ActivityType.elliptical: 1.0,
-  };
-
   static final _preferencesSpecsTemplate = [
     MetricSpec(
       metric: metrics[0],
@@ -56,27 +39,27 @@ class MetricSpec {
       unit: "W",
       thresholdTagPostfix: thresholdPrefix + metrics[0],
       oldThresholdDefaultInts: {
-        sportPrefixes[0]: 300,
-        sportPrefixes[1]: 180,
-        sportPrefixes[2]: 100,
-        sportPrefixes[3]: 100,
-        sportPrefixes[4]: 180,
+        SportSpec.sportPrefixes[0]: 300,
+        SportSpec.sportPrefixes[1]: 180,
+        SportSpec.sportPrefixes[2]: 100,
+        SportSpec.sportPrefixes[3]: 100,
+        SportSpec.sportPrefixes[4]: 180,
       },
       thresholdDefaultInts: {
-        sportPrefixes[0]: 320,
-        sportPrefixes[1]: 180,
-        sportPrefixes[2]: 100,
-        sportPrefixes[3]: 100,
-        sportPrefixes[4]: 180,
+        SportSpec.sportPrefixes[0]: 320,
+        SportSpec.sportPrefixes[1]: 180,
+        SportSpec.sportPrefixes[2]: 100,
+        SportSpec.sportPrefixes[3]: 100,
+        SportSpec.sportPrefixes[4]: 180,
       },
       zonesTagPostfix: metrics[0] + zonesPostfix,
       oldZoneDefaultInts: [55, 75, 90, 105, 120, 150],
       zonesDefaultInts: {
-        sportPrefixes[0]: [55, 75, 90, 105, 120, 150],
-        sportPrefixes[1]: [55, 75, 90, 105, 120, 150],
-        sportPrefixes[2]: [55, 75, 90, 105, 120, 150],
-        sportPrefixes[3]: [55, 75, 90, 105, 120, 150],
-        sportPrefixes[4]: [55, 75, 90, 105, 120, 150],
+        SportSpec.sportPrefixes[0]: [55, 75, 90, 105, 120, 150],
+        SportSpec.sportPrefixes[1]: [55, 75, 90, 105, 120, 150],
+        SportSpec.sportPrefixes[2]: [55, 75, 90, 105, 120, 150],
+        SportSpec.sportPrefixes[3]: [55, 75, 90, 105, 120, 150],
+        SportSpec.sportPrefixes[4]: [55, 75, 90, 105, 120, 150],
       },
       icon: Icons.bolt,
       indexDisplayDefault: false,
@@ -87,27 +70,27 @@ class MetricSpec {
       unit: "mph",
       thresholdTagPostfix: thresholdPrefix + metrics[1],
       oldThresholdDefaultInts: {
-        sportPrefixes[0]: 32,
-        sportPrefixes[1]: 16,
-        sportPrefixes[2]: 7,
-        sportPrefixes[3]: 1,
-        sportPrefixes[4]: 7,
+        SportSpec.sportPrefixes[0]: 32,
+        SportSpec.sportPrefixes[1]: 16,
+        SportSpec.sportPrefixes[2]: 7,
+        SportSpec.sportPrefixes[3]: 1,
+        SportSpec.sportPrefixes[4]: 7,
       },
       thresholdDefaultInts: {
-        sportPrefixes[0]: 42,
-        sportPrefixes[1]: 14,
-        sportPrefixes[2]: 7,
-        sportPrefixes[3]: 1,
-        sportPrefixes[4]: 7,
+        SportSpec.sportPrefixes[0]: 42,
+        SportSpec.sportPrefixes[1]: 14,
+        SportSpec.sportPrefixes[2]: 7,
+        SportSpec.sportPrefixes[3]: 1,
+        SportSpec.sportPrefixes[4]: 7,
       },
       zonesTagPostfix: metrics[1] + zonesPostfix,
       oldZoneDefaultInts: [55, 75, 90, 105, 120, 150],
       zonesDefaultInts: {
-        sportPrefixes[0]: [77, 87, 94, 100, 104, 112],
-        sportPrefixes[1]: [77, 87, 94, 100, 104, 112],
-        sportPrefixes[2]: [77, 87, 94, 100, 104, 112],
-        sportPrefixes[3]: [77, 87, 94, 100, 104, 112],
-        sportPrefixes[4]: [77, 87, 94, 100, 104, 112],
+        SportSpec.sportPrefixes[0]: [77, 87, 94, 100, 104, 112],
+        SportSpec.sportPrefixes[1]: [77, 87, 94, 100, 104, 112],
+        SportSpec.sportPrefixes[2]: [77, 87, 94, 100, 104, 112],
+        SportSpec.sportPrefixes[3]: [77, 87, 94, 100, 104, 112],
+        SportSpec.sportPrefixes[4]: [77, 87, 94, 100, 104, 112],
       },
       icon: Icons.speed,
       indexDisplayDefault: false,
@@ -118,27 +101,27 @@ class MetricSpec {
       unit: "rpm",
       thresholdTagPostfix: thresholdPrefix + metrics[2],
       oldThresholdDefaultInts: {
-        sportPrefixes[0]: 100,
-        sportPrefixes[1]: 180,
-        sportPrefixes[2]: 90,
-        sportPrefixes[3]: 90,
-        sportPrefixes[4]: 90,
+        SportSpec.sportPrefixes[0]: 100,
+        SportSpec.sportPrefixes[1]: 180,
+        SportSpec.sportPrefixes[2]: 90,
+        SportSpec.sportPrefixes[3]: 90,
+        SportSpec.sportPrefixes[4]: 90,
       },
       thresholdDefaultInts: {
-        sportPrefixes[0]: 100,
-        sportPrefixes[1]: 170,
-        sportPrefixes[2]: 85,
-        sportPrefixes[3]: 80,
-        sportPrefixes[4]: 150,
+        SportSpec.sportPrefixes[0]: 100,
+        SportSpec.sportPrefixes[1]: 170,
+        SportSpec.sportPrefixes[2]: 85,
+        SportSpec.sportPrefixes[3]: 80,
+        SportSpec.sportPrefixes[4]: 150,
       },
       zonesTagPostfix: metrics[2] + zonesPostfix,
       oldZoneDefaultInts: [25, 37, 50, 75, 100, 120],
       zonesDefaultInts: {
-        sportPrefixes[0]: [25, 37, 50, 75, 100, 120],
-        sportPrefixes[1]: [77, 87, 94, 100, 104, 112],
-        sportPrefixes[2]: [77, 87, 94, 100, 104, 112],
-        sportPrefixes[3]: [77, 87, 94, 100, 104, 112],
-        sportPrefixes[4]: [77, 87, 94, 100, 104, 112],
+        SportSpec.sportPrefixes[0]: [25, 37, 50, 75, 100, 120],
+        SportSpec.sportPrefixes[1]: [77, 87, 94, 100, 104, 112],
+        SportSpec.sportPrefixes[2]: [77, 87, 94, 100, 104, 112],
+        SportSpec.sportPrefixes[3]: [77, 87, 94, 100, 104, 112],
+        SportSpec.sportPrefixes[4]: [77, 87, 94, 100, 104, 112],
       },
       icon: Icons.directions_bike,
       indexDisplayDefault: false,
@@ -149,27 +132,27 @@ class MetricSpec {
       unit: "bpm",
       thresholdTagPostfix: thresholdPrefix + metrics[3],
       oldThresholdDefaultInts: {
-        sportPrefixes[0]: 180,
-        sportPrefixes[1]: 180,
-        sportPrefixes[2]: 180,
-        sportPrefixes[3]: 180,
-        sportPrefixes[4]: 180,
+        SportSpec.sportPrefixes[0]: 180,
+        SportSpec.sportPrefixes[1]: 180,
+        SportSpec.sportPrefixes[2]: 180,
+        SportSpec.sportPrefixes[3]: 180,
+        SportSpec.sportPrefixes[4]: 180,
       },
       thresholdDefaultInts: {
-        sportPrefixes[0]: 153,
-        sportPrefixes[1]: 153,
-        sportPrefixes[2]: 153,
-        sportPrefixes[3]: 153,
-        sportPrefixes[4]: 153,
+        SportSpec.sportPrefixes[0]: 153,
+        SportSpec.sportPrefixes[1]: 153,
+        SportSpec.sportPrefixes[2]: 153,
+        SportSpec.sportPrefixes[3]: 153,
+        SportSpec.sportPrefixes[4]: 153,
       },
       zonesTagPostfix: metrics[3] + zonesPostfix,
       oldZoneDefaultInts: [50, 60, 70, 80, 90, 100],
       zonesDefaultInts: {
-        sportPrefixes[0]: [80, 88, 92, 152, 156, 163],
-        sportPrefixes[1]: [80, 88, 92, 152, 156, 163],
-        sportPrefixes[2]: [80, 88, 92, 152, 156, 163],
-        sportPrefixes[3]: [80, 88, 92, 152, 156, 163],
-        sportPrefixes[4]: [80, 88, 92, 152, 156, 163],
+        SportSpec.sportPrefixes[0]: [80, 88, 92, 152, 156, 163],
+        SportSpec.sportPrefixes[1]: [80, 88, 92, 152, 156, 163],
+        SportSpec.sportPrefixes[2]: [80, 88, 92, 152, 156, 163],
+        SportSpec.sportPrefixes[3]: [80, 88, 92, 152, 156, 163],
+        SportSpec.sportPrefixes[4]: [80, 88, 92, 152, 156, 163],
       },
       icon: Icons.favorite,
       indexDisplayDefault: false,
@@ -226,20 +209,12 @@ class MetricSpec {
   String get zoneIndexDescription =>
       "$zoneIndexDisplayDescriptionPart1 $title $zoneIndexDisplayDescriptionPart2";
 
-  static String sport2Sport(String sport) {
-    return sport == ActivityType.kayaking ||
-            sport == ActivityType.canoeing ||
-            sport == ActivityType.rowing
-        ? paddleSport
-        : sport;
-  }
-
   String oldThresholdDefault(String sport) {
-    return oldThresholdDefaultInts[sport2Sport(sport)].toString();
+    return oldThresholdDefaultInts[SportSpec.sport2Sport(sport)].toString();
   }
 
   String thresholdDefault(String sport) {
-    return thresholdDefaultInts[sport2Sport(sport)].toString();
+    return thresholdDefaultInts[SportSpec.sport2Sport(sport)].toString();
   }
 
   String intArrayToString(List<int> intArray) {
@@ -251,19 +226,15 @@ class MetricSpec {
   }
 
   String zonesDefault(String sport) {
-    return intArrayToString(zonesDefaultInts[sport2Sport(sport)]!);
+    return intArrayToString(zonesDefaultInts[SportSpec.sport2Sport(sport)]!);
   }
 
   String thresholdTag(String sport) {
-    return sport2Sport(sport) + "_" + thresholdTagPostfix;
+    return SportSpec.sport2Sport(sport) + "_" + thresholdTagPostfix;
   }
 
   String zonesTag(String sport) {
-    return sport2Sport(sport) + "_" + zonesTagPostfix;
-  }
-
-  static String slowSpeedTag(String sport) {
-    return slowSpeedTagPrefix + sport2Sport(sport);
+    return SportSpec.sport2Sport(sport) + "_" + zonesTagPostfix;
   }
 
   void updateMultiLineUnit() {