+++ lib/preferences/show_pacer.dart
@@ -1,0 +1,10 @@
+const showPacer = "Show Pacer Feature";
+const showPacerTag = "show_pacer";
+const showPacerDefault = false;
+const showPacerDescription =
+    "On: A pacer with the given speed can be shown on the track and leaderboard. "
+    "Off: No pacer is shown. Note: Pacer speed can be configured similarly as slow speed "
+    "per sport within the Zone Preferences";
+
+const pacerColor = 0xFF000000; // Black
+const pacerText = "P";